Complete work on basis vector sign optimization in BasisTransform

+++ BasisTransform.h
@@ -95,7 +95,7 @@ protected:
         cv::InputArray inputVectors, cv::OutputArray outputVectors, const bool &useMean = false,
         const VectorDirection &basisVecDir = VectorDirection::COLUMNVECTORS);
     ///Randomly choose numberOfPixels rows from sourcePixels, copy them to the subsample OutputArray.
-    void CreatePixelSubsample(cv::InputArray sourcePixels, cv::OutputArray subsample, const long int &numberOfPixels);
+    void CreatePixelSubsample(cv::InputArray sourcePixels, cv::OutputArray subsample, const int &numberOfPixels);
     ///Set the member basis vectors. Second parameter is direction of input vectors. Store basis vectors as row vectors.
     void SetBasisVectors(cv::InputArray basisVectors, const VectorDirection &vecDir = VectorDirection::ROWVECTORS);
     ///Set the member point mean
